Added .ino File
Created the Arduino .ino file and moved the ISR from the Step sensor.cpp to the .ino.
Also Added the volatile keyword and included the sei() function to the ISR. The sei() function ensures that interrupts are also interruptable which is key to the whole step sensor.

Currently working on other bugs, not sure what the problem is (I suspect it's the physical PIR sensor modules) but the code step direction isn't always right :) .